chore: refactor common types out to separate package
