+++ packages/asset-service/package.json
@@ -22,7 +22,8 @@
     "type-check": "tsc --project ./tsconfig.json --noEmit"
   },
   "dependencies": {
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "@shapeshiftoss/types": "^1.0.0"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.172"
