--- packages/asset-service/src/types.ts
@@ -1,44 +1,0 @@
-type AbstractAsset = {
-  chain: ChainTypes
-  network: NetworkTypes
-  symbol: string
-  name: string
-  precision: number
-  slip44: number
-  color: string
-  secondaryColor: string
-  icon: string
-  explorer: string
-  explorerTxLink: string
-  sendSupport: boolean
-  receiveSupport: boolean
-}
-
-type OmittedTokenAssetFields = 'chain' | 'network' | 'slip44' | 'explorer' | 'explorerTxLink'
-type TokenAssetFields = {
-  tokenId: string
-  contractType: ContractTypes
-}
-export type TokenAsset = Omit<AbstractAsset, OmittedTokenAssetFields> & TokenAssetFields
-export type BaseAsset = AbstractAsset & { tokens?: TokenAsset[] }
-export type Asset = AbstractAsset & Partial<TokenAssetFields>
-
-export enum ContractTypes {
-  ERC20 = 'ERC20',
-  ERC721 = 'ERC721',
-  OTHER = 'OTHER',
-  NONE = 'NONE'
-}
-
-export enum ChainTypes {
-  Ethereum = 'ethereum',
-  Bitcoin = 'bitcoin',
-  Litecoin = 'litecoin'
-}
-
-export enum NetworkTypes {
-  MAINNET = 'MAINNET',
-  TESTNET = 'TESTNET', // BTC, lTC, etc...
-  ETH_ROPSTEN = 'ETH_ROPSTEN',
-  ETH_RINKEBY = 'ETH_RINKEBY'
-}
