+++ packages/market-service/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@shapeshiftoss/market-service",
   "version": "1.0.0",
-  "description": "Package to share reusable code",
+  "description": "ShapeShift market data service",
   "homepage": "",
   "license": "MIT",
   "main": "./dist/index.js",
@@ -21,7 +21,7 @@
     "type-check": "tsc --project ./tsconfig.json --noEmit"
   },
   "dependencies": {
-    "@shapeshiftoss/asset-service": "^1.0.0",
+    "@shapeshiftoss/types": "^1.0.0",
     "axios": "^0.21.2",
     "dayjs": "^1.10.6"
   },
