+++ packages/market-service/src/api.ts
@@ -1,49 +1,7 @@
-export enum ChainTypes {
-  Ethereum = 'ethereum',
-  Bitcoin = 'bitcoin',
-  Litecoin = 'litecoin'
-}
-
-export type MarketData = {
-  price: string
-  marketCap: string
-  volume: string
-  changePercent24Hr: number
-}
-
-export enum HistoryTimeframe {
-  HOUR = '1H',
-  DAY = '24H',
-  WEEK = '1W',
-  MONTH = '1M',
-  YEAR = '1Y',
-  ALL = 'All'
-}
-
-export type HistoryData = {
-  price: number
-  date: string
-}
-
-export type PriceHistoryArgs = {
-  chain: ChainTypes
-  timeframe: HistoryTimeframe
-  tokenId?: string
-}
-
-export type MarketDataArgs = {
-  chain: ChainTypes
-  tokenId?: string
-}
-
-export type MarketDataType = (args: MarketDataArgs) => Promise<MarketData | null>
-
-export type PriceHistoryType = (args: PriceHistoryArgs) => Promise<HistoryData[] | null>
+import { MarketDataType, PriceHistoryType } from '@shapeshiftoss/types'
 
 export interface MarketService {
   baseUrl: string
-
   getMarketData: MarketDataType
-
   getPriceHistory: PriceHistoryType
 }
