+++ packages/swapper/src/swappers/zrx/ZrxSwapper.ts
@@ -1,7 +1,14 @@
 import Web3 from 'web3'
-import { Asset, ChainTypes } from '@shapeshiftoss/asset-service'
+import {
+  Asset,
+  BuildQuoteTxArgs,
+  ChainTypes,
+  GetQuoteInput,
+  Quote,
+  SwapperType
+} from '@shapeshiftoss/types'
 import { ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
-import { BuildQuoteTxArgs, GetQuoteInput, Quote, Swapper, SwapperType } from '../../api'
+import { Swapper } from '../../api'
 
 import { buildQuoteTx } from './buildQuoteTx/buildQuoteTx'
 import { getZrxQuote } from './getQuote/getQuote'
