+++ packages/swapper/src/swappers/zrx/getQuote/getQuote.ts
@@ -1,7 +1,7 @@
-import { ChainTypes } from '@shapeshiftoss/asset-service'
 import { AxiosResponse } from 'axios'
 import BigNumber from 'bignumber.js'
-import { GetQuoteInput, Quote, QuoteResponse, SwapSource } from '../../..'
+import { ChainTypes, GetQuoteInput, Quote, QuoteResponse, SwapSource } from '@shapeshiftoss/types'
+
 import { MAX_ZRX_TRADE, APPROVAL_GAS_LIMIT, DEFAULT_SOURCE } from '../utils/constants'
 import { zrxService } from '../utils/zrxService'
 import { normalizeAmount } from '../utils/helpers/helpers'
