+++ packages/types/README.md
@@ -1,0 +1,13 @@
+## ShapeShift Types
+
+## Getting Started
+
+```sh
+yarn add @shapeshiftoss/types
+```
+
+## Usage
+
+```ts
+import { Asset, ChainTypes, Network } from '@shapeshiftoss/types'
+```
